+++ src/generator/with_rand08.rs
@@ -19,29 +19,7 @@ impl<T: RngCore> Scru128Generator<Adapter<T>> {
     /// Creates a generator object with a specified random number generator that implements
     /// [`RngCore`] from `rand` (v0.8) crate. The specified random number generator should be
     /// cryptographically strong and securely seeded.
-    ///
-    /// # Examples
-    ///
-    /// ```ignore
-    /// # #[cfg(feature = "default_rng")]
-    /// # {
-    /// # use rand08 as rand;
-    /// use scru128::Scru128Generator;
-    ///
-    /// let mut g = Scru128Generator::with_rand08(rand::thread_rng());
-    /// println!("{}", g.generate());
-    /// # }
-    /// ```
     pub const fn with_rand08(rng: T) -> Self {
         Self::with_rng(Adapter(rng))
-    }
-}
-
-/// This is a deprecated blanket impl retained for backward compatibility. Do not depend on this
-/// impl; use [`Scru128Generator::with_rand08()`] instead.
-#[cfg(feature = "rand")]
-impl<T: RngCore> Scru128Rng for T {
-    fn next_u32(&mut self) -> u32 {
-        self.next_u32()
     }
 }